Create user domain and review domain

+++ src/main/java/com/example/demo/src/user/model/PostLoginRes.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public class PostLoginRes {
 
-    private int userIdx;
+    private long userIdx;
     private String jwt;
 }
